feat(analysis): enhance stock listing functionality with query support
- Added a new Query field to StockListRequest for fuzzy searching by name, ts_code, or symbol.
- Updated ListStocks method to incorporate query filtering in SQL statements.
- Enhanced HTTP handler to accept query parameter for stock listing requests.
- Improved documentation to reflect the new query functionality in API descriptions.

Made-with: Cursor

+++ qdhub/internal/infrastructure/persistence/repository/quant_data_store_repo.go
@@ -284,7 +284,7 @@ func (r *DataTypeMappingRuleRepositoryImpl) InitDefaultRules() error {
 	defaultRules := []*datastore.DataTypeMappingRule{
 		datastore.NewDataTypeMappingRule("tushare", "str", "duckdb", "VARCHAR", 100, true),
 		datastore.NewDataTypeMappingRule("tushare", "float", "duckdb", "DOUBLE", 100, true),
-		datastore.NewDataTypeMappingRule("tushare", "int", "duckdb", "INTEGER", 100, true),
+		datastore.NewDataTypeMappingRule("tushare", "int", "duckdb", "BIGINT", 100, true),
 		datastore.NewDataTypeMappingRule("tushare", "datetime", "duckdb", "TIMESTAMP", 100, true),
 		datastore.NewDataTypeMappingRule("tushare", "date", "duckdb", "DATE", 100, true),
 	}
